pyproject.toml
[tool.poetry]
name = "auth-system"
version = "0.1.0"
description = ""
authors = ["awsconsultant9"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["email"], version = "^0.116.1"}
pydantic = {extras = ["email"], version = "^2.11.7"}
sqlalchemy = "^2.0.42"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}


[tool.poetry.group.dev.dependencies]
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
